+++ src/utils/Vec_ut.cpp
@@ -165,7 +165,7 @@ static void VecTest()
 
     {
         str::Str<char> v(0);
-        for (int i = 0; i < 32; i++) {
+        for (size_t i = 0; i < 32; i++) {
             assert(v.Count() == i * 6);
             v.Append("lambd", 5);
             if (i % 2 == 0)
@@ -174,7 +174,7 @@ static void VecTest()
                 v.Push('a');
         }
 
-        for (int i=1; i<=16; i++) {
+        for (size_t i=1; i<=16; i++) {
             v.RemoveAt((16 - i) * 6, 6);
             assert(v.Count() == (32 - i) * 6);
         }
